[#120] Backup vem_bold file and replace it by new module with new function. Fix convergence criteria
